Software Defined PLL working as expected

+++ Software_Defined_PLL/main.h
@@ -1,7 +1,7 @@
 #ifndef MAIN_H
 #define MAIN_H
 
-float average_shift(float received_signal, float local_osc, float previous_signal);
-float lowPass_filter(float mixed_signal, float previous_signal);
+float average_shift(double received_signal, double local_osc, double previous_signal);
+float lowPass_filter(double mixed_signal, double previous_signal);
 
 #endif
